pyproject.toml
[project]
name = "anymate-saas"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cookiecutter>=2.6.0",
    "django>=5.2.5",
    "django-cors-headers>=4.7.0",
    "django-environ>=0.12.0",
    "django-filter>=25.1",
    "django-tailwind[reload]>=4.2.0",
    "django-userforeignkey>=0.5.0",
    "djangorestframework>=3.16.1",
    "drf-nested-routers>=0.94.2",
    "drf-spectacular>=0.28.0",
    "markdown>=3.8.2",
    "psycopg2>=2.9.10",
]
